Add an AR cude, add the Cube Manager script
Add AR Cube is a cube which you can put into the world in which you track a plane.
The Cube Manager script created to check when we trach a palne then we can put there a cube. If I touch a cube which I have put at one point the the cube was deleted.